add hook `useVegaSignalEmbed`
